// Source/OregonTilth.Web/src/app/shared/models/forms/crop-yield-information/crop-yield-information-create-dto.ts
export class CropYieldInformationCreateDto {
    WorkbookID: number;
    CropID: number;
    CropUnitID: number;
    HarvestedYieldPerStandardUnitOfSpace: number;
    MarketableYieldPerStandardUnitOfSpace: number;
    PackagingCostPerCropUnit: number;
    PricePerCropUnit: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}